+++ spec/javascripts/govuk_publishing_components/analytics-ga4/ga4-event-tracker.spec.js
@@ -462,4 +462,51 @@ describe('Google Analytics event tracking', function () {
       expect(window.dataLayer[0]).toEqual(expected)
     })
   })
+
+  describe('doing tracking on an data-ga4-expandable element', function () {
+    beforeEach(function () {
+      element.innerHTML = '<header data-module="ga4-event-tracker" data-ga4-expandable><nav>' +
+      '<button aria-expanded="false">Menu</button>' +
+      '<ul><li><button aria-expanded="false"">Topics</button></li>' +
+      '<li><button aria-expanded="false">Government activity</button></li></ul>' +
+      '<button aria-expanded="false">Search</button>' +
+      '</nav></header>'
+      document.body.appendChild(element)
+      new GOVUK.Modules.Ga4EventTracker(element).init()
+    })
+
+    it('tracks expanding/collapsing sections under the data-ga4-expandable attribute', function () {
+      var expected = new GOVUK.analyticsGa4.Schemas().eventSchema()
+      expected.event = 'event_data'
+      expected.event_data.event_name = 'select_content'
+      expected.event_data.type = 'header menu bar'
+      expected.govuk_gem_version = 'aVersion'
+      var buttons = document.querySelectorAll('button')
+      expected.event_data.index_total = buttons.length
+
+      for (var i = 0; i < buttons.length; i++) {
+        window.dataLayer = []
+        var button = buttons[i]
+        button.setAttribute('data-ga4', JSON.stringify({
+          event_name: 'select_content',
+          type: 'header menu bar',
+          text: button.textContent,
+          section: button.textContent,
+          index: i + 1,
+          index_total: buttons.length
+        }))
+
+        button.click()
+        expected.event_data.action = 'opened'
+        expected.event_data.text = button.textContent
+        expected.event_data.section = button.textContent
+        expected.event_data.index = i + 1
+        expect(window.dataLayer[0]).toEqual(expected)
+        button.setAttribute('aria-expanded', 'true')
+        expected.event_data.action = 'closed'
+        button.click()
+        expect(window.dataLayer[1]).toEqual(expected)
+      }
+    })
+  })
 })